+++ Content/Python/UEFN_Toolbelt/tools/__init__.py
@@ -73,3 +73,4 @@ from . import geometry_tools
 from . import movie_render_tools
 from . import viewport_tools
 from . import actor_visibility
+from . import verse_templates